+++ code/qcommon/qcommon.h
@@ -753,13 +753,13 @@ temp file loading
 #define Z_TagMalloc(size, tag) Z_TagMallocDebug(size, tag, #size, __FILE__, __LINE__)
 #define Z_Malloc(size) Z_MallocDebug(size, #size, __FILE__, __LINE__)
 #define S_Malloc(size) S_MallocDebug(size, #size, __FILE__, __LINE__)
-void* Z_TagMallocDebug(int size, int tag, char* label, char* file, int line); // NOT 0 filled memory
-void* Z_MallocDebug(int size, char* label, char* file, int line); // returns 0 filled memory
-void* S_MallocDebug(int size, char* label, char* file, int line); // returns 0 filled memory
+void* Z_TagMallocDebug(size_t size, int tag, char* label, char* file, int line); // NOT 0 filled memory
+void* Z_MallocDebug(size_t size, char* label, char* file, int line); // returns 0 filled memory
+void* S_MallocDebug(size_t size, char* label, char* file, int line); // returns 0 filled memory
 #else
-void* Z_TagMalloc(int size, int tag); // NOT 0 filled memory
-void* Z_Malloc(int size); // returns 0 filled memory
-void* S_Malloc(int size); // NOT 0 filled memory only for small allocations
+void* Z_TagMalloc(size_t size, int tag); // NOT 0 filled memory
+void* Z_Malloc(size_t size); // returns 0 filled memory
+void* S_Malloc(size_t size); // NOT 0 filled memory only for small allocations
 #endif
 void Z_Free(void* ptr);
 int Z_AvailableMemory(void);
